Ticket #1740
Message: Fixing the VS 2010 build. As part of this I converted the Contrib Solution and its projects to VS 2010.

Owners Code Review: No code was changed, just new solutions and projects created, but I did talk to Ben Marchant and Smanders about the new solution and projects and the new library name and they concurred.

git-svn-id: https://isrsvn.usurf.usu.edu/svn/isr/roots/Contrib/jasper@17024 a06e8daa-702c-0410-a579-eb5e6c4fa09d